refactor: alter MD5 to random string 5 length
