+++ backend/src/main/java/study/memoravel/util/Encoding.java
@@ -2,6 +2,9 @@ package study.memoravel.util;
 
 import org.springframework.security.crypto.bcrypt.BCrypt;
 
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
 public class Encoding {
 
     public static String getBCrypt(String password) {
@@ -10,5 +13,16 @@ public class Encoding {
 
     public static Boolean checkBCrypt(String value, String target) {
         return BCrypt.checkpw(value, target);
+    }
+
+    public static String getMD5(String email) throws NoSuchAlgorithmException {
+        MessageDigest md = MessageDigest.getInstance("MD5");
+        md.update(email.getBytes());
+        byte[] byteData = md.digest();
+        StringBuilder sb = new StringBuilder();
+        for (byte byteDatum : byteData) {
+            sb.append(Integer.toString((byteDatum & 0xff) + 0x100, 16).substring(1));
+        }
+        return sb.toString();
     }
 }
