--- backend/src/main/java/study/memoravel/util/MD5.java
@@ -1,17 +1,0 @@
-package study.memoravel.util;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class MD5 {
-    public static String getMD5(String email) throws NoSuchAlgorithmException {
-        MessageDigest md = MessageDigest.getInstance("MD5");
-        md.update(email.getBytes());
-        byte[] byteData = md.digest();
-        StringBuilder sb = new StringBuilder();
-        for (byte byteDatum : byteData) {
-            sb.append(Integer.toString((byteDatum & 0xff) + 0x100, 16).substring(1));
-        }
-        return sb.toString();
-    }
-}
